requirements.txt
Flask~=1.1
gunicorn==20.0.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
simplejson==3.17.2
Werkzeug==1.0.1
WTForms==2.3.3
psycopg2-binary==2.8.6
boto3==1.16.18
